+++ src/main/java/org/uom/lefterisxris/codetour/tours/actions/TourStepGeneratorAction.java
@@ -1,18 +1,17 @@
 package org.uom.lefterisxris.codetour.tours.actions;
 
-import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.Messages;
 import com.intellij.openapi.vfs.VirtualFile;
 import org.jetbrains.annotations.NotNull;
 import org.uom.lefterisxris.codetour.tours.domain.Step;
 import org.uom.lefterisxris.codetour.tours.domain.Tour;
 import org.uom.lefterisxris.codetour.tours.state.StateManager;
 import org.uom.lefterisxris.codetour.tours.state.TourUpdateNotifier;
+import org.uom.lefterisxris.codetour.tours.ui.StepDescriptionEditor;
 import org.uom.lefterisxris.codetour.tours.ui.TourSelectionDialogWrapper;
 
 import java.util.Optional;
@@ -65,13 +64,11 @@ public class TourStepGeneratorAction extends AnAction {
          final Step step = generateStep(virtualFile, line);
 
          // Provide a dialog for step's description
-         final String updatedDescription = Messages.showMultilineInputDialog(project,
-               String.format("Edit Step's '%s' description", step.getTitle()),
-               "Edit Step",
-               step.getDescription(),
-               AllIcons.Actions.Edit, null);
-         if (updatedDescription != null)
-            step.setDescription(updatedDescription);
+         final String updatedDescription = StepDescriptionEditor.show(project, step, true);
+         if (updatedDescription == null)
+            return; // i.e. cancel the step creation
+
+         step.setDescription(updatedDescription);
 
          activeTour.get().getSteps().add(step);
          stateManager.updateTour(activeTour.get());