+++ CHANGELOG.md
@@ -4,8 +4,17 @@
 
 ## [Unreleased]
 
+### Added
+
+- New Icons
+- New Step's Description default value is pre-selected for easier user input
+- Added markdown support
+- Sample Tours (for demonstration)
+
 ## [0.0.2]
+
 ### Implemented
+
 - Load from json, Show In Tool Pane, Navigate with click
 - Navigate with Previous/Next buttons on Tool Pane
 - Create|Delete Tour from Tool Pane